// src/model/interfaces/empresa.interfaces.ts
export default interface Empresa {
    idEmpresa: number
    celular: string
    certificadoClave: string
    certificadoRuta: string
    claveApiSunat: string
    claveSol: string
    condiciones: string
    domicilio: string
    email: string
    giroComercial: number
    idApiSunat: string
    image: number[]
    imagenRuta: string
    nombre: string
    nombreComercial: string
    numeroDocumento: string
    paginaWeb: string
    razonSocial: string
    telefono: string
    terminos: string
    tipoDocumento: number
    ubigeo: number
    usuarioSol: string
  }